Viddle
Viddle is a riddle game hosted on a website.

Your goal is to guess the answer to a riddle based on a question and clues.

If you think the riddle is too hard to solve, you can click the "hint" button. The button will provide you with an image containing a visual clue to the answer.

To play the game press the "play" button.

If you have questions about the rules of the game, you can click on "How to Play" and the page will provide information about the game.

AI use: CS50 AI duck and Copilot were used in the making of the website.

Copilot AI was used in creating the riddle questions only, it was not used for writing code.

The CS50 AI Duck was used to provide syntax information, correct syntax and logic errors, and logic development for some functionalities.

The CS50 AI Duck was used to write the majority of CSS styling.

The CS50 AI Duck was not used in writing any Javascript code, only provided some logic and syntax for writing.

Functionalities include: The logic/idea of using a second dictionary to search for a matching key in the riddle dictionary. Some event listener functions. CSS styling.

All syntax used in Javascript, HTML, and CSS was provided by the CS50 AI Duck.